# Support `typing.Final` annotations