@jcrist 2023-03-25
A quick benchmarks trying to measure just the type annotation parsing overhead. For most usage patterns the type annotation parsing happens only once and is cached for subsequent usage, so I wouldn't expect this perf increase to matter to most users. Still, speeding this up can decrease startup times.

**Main Branch**

```python
In [1]: import msgspec

In [2]: typ = dict[str, list[int]]

In [3]: msg = b'{"x":[1]}'

In [4]: %timeit msgspec.json.decode(msg, type=typ)
1.78 µs ± 18.2 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

**This PR**

```python
In [1]: import msgspec

In [2]: typ = dict[str, list[int]]

In [3]: msg = b'{"x":[1]}'

In [4]: %timeit msgspec.json.decode(msg, type=typ)
653 ns ± 5.72 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```